Fix default preference value for alpha slider & refactoring
