+++ app/src/main/res/values/strings.xml
@@ -9,7 +9,7 @@
  Release: Working copy
  Locale: en, English
  Exported by: ChilliBits
- Exported at: Sun, 21 Feb 2021 22:29:54 +0000
+ Exported at: Tue, 23 Feb 2021 03:47:20 +0000
 -->
 <resources xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2">
     <string name="ok">OK</string>
